requirements.txt
Django==1.3
South==0.7.3
django-annoying==0.7.6
django-extensions==0.6
django-indexer==0.3.0
django-paging==0.2.4
django-registration==0.7
django-sentry==1.8.6.2
django-staticfiles==1.1.2
django-tagging==0.3.1
psycopg2==2.4.1
Markdown==2.1.1
PIL==1.1.7
